Reworked the flag object and changed how terrain dimensions are calculated
All the calculations are done in base res and the scaling is only applied right before drawing now
so the flag and hitbox will automatically be placed on where the hole is generated in the match in the proper location and scale

+++ src/flag.h
@@ -14,7 +14,7 @@ private:
 
 public:
     Flag();
-    Flag(int baseX, int baseY, float scale = 1.0f);
+    Flag(float baseX, float baseY, float scale = 1.0f);
     // ~Flag();
 
     void draw() const;    //draw the flag
